+++ crates/fandwill-vo/src/resources.rs
@@ -1,9 +1,9 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use url::Url;
-use utoipa::ToSchema;
 
-#[derive(Debug, ToSchema, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 pub struct ResourceVO {
     pub id: String,
     pub mime_type: Option<String>,
@@ -12,7 +12,8 @@ pub struct ResourceVO {
     pub created_at: DateTime<Utc>,
 }
 
-#[derive(Debug, ToSchema, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 pub struct CreateResourceVO {
     pub id: String,
     pub upload_url: Url,